+++ toml-config/src/lib.rs
@@ -30,10 +30,10 @@ use {
         switch_to_vt,
         theme::{reset_colors, reset_font, reset_sizes, set_font},
         video::{
-            Connector, DrmDevice, connectors, drm_devices, on_connector_connected,
-            on_connector_disconnected, on_graphics_initialized, on_new_connector,
-            on_new_drm_device, set_direct_scanout_enabled, set_gfx_api, set_tearing_mode,
-            set_vrr_cursor_hz, set_vrr_mode,
+            ColorSpace, Connector, DrmDevice, TransferFunction, connectors, drm_devices,
+            on_connector_connected, on_connector_disconnected, on_graphics_initialized,
+            on_new_connector, on_new_drm_device, set_direct_scanout_enabled, set_gfx_api,
+            set_tearing_mode, set_vrr_cursor_hz, set_vrr_mode,
         },
         xwayland::set_x_scaling_mode,
     },
@@ -587,6 +587,11 @@ impl Output {
         }
         if let Some(format) = self.format {
             c.set_format(format);
+        }
+        if self.color_space.is_some() || self.transfer_function.is_some() {
+            let cs = self.color_space.unwrap_or(ColorSpace::DEFAULT);
+            let tf = self.transfer_function.unwrap_or(TransferFunction::DEFAULT);
+            c.set_colors(cs, tf);
         }
     }
 }